I added a :gem_name resource property
This commit augments the "_mariadb" and "_mysql" resources with a
property named "gem_name", whose default is "mysql2". This property
allows one to specify an alternate gem name in situations where
"mysql2" cannot be used. It applies in particular to use cases where
the deployment environment lacks a C compiler and one must pre-compile
binary extensions into a gem with another name.

Change-Id: I93ffb817fc243ac0243625a9d1d2e7cbd69fee5d